move consistent hashring balancer into balancer pkg
